Add RouterSource typings and remove typing dependency to switch-path
