--- src/custom-typings/main.d.ts
@@ -1,1 +1,0 @@
-/// <reference path="./switch-path.d.ts" />
