+++ src/interfaces.ts
@@ -1,0 +1,8 @@
+export interface RouteDefinitions {
+  [sourcePath: string]: any;
+}
+
+export interface SwitchPathReturn {
+  path: string;
+  value: any;
+}
