+++ src/switch-path.d.ts
@@ -1,0 +1,3 @@
+declare module 'switch-path' {
+  export default function switchPath(sourcePath: string, routes: {}): {};
+}
