+++ typings.json
@@ -1,8 +1,5 @@
 {
   "name": "cyclic-router",
-  "dependencies": {
-    "@cycle/history": "npm:@cycle/history"
-  },
   "ambientDevDependencies": {
     "mocha": "github:DefinitelyTyped/DefinitelyTyped/mocha/mocha.d.ts#d6dd320291705694ba8e1a79497a908e9f5e6617",
     "node": "github:DefinitelyTyped/DefinitelyTyped/node/node.d.ts#20e1eb9616922d382d918cc5a21870a9dbe255f5"
